Feat: API for Label
