+++ API/api/api_v1/endpoints/adapter_predict.py
@@ -39,44 +39,3 @@ def text_label(data: text_label_body):
         })
     return out_Tokens
 
-
-
-@router.post("/model/template", tags = ["Predict"], status_code=status.HTTP_200_OK)
-def recommand_template(data: text_label_body):
-    return {
-        "predict": [
-            {
-                "name": "Full Payment Upon Demand",
-                "confidence": 0.9
-            },
-            {
-                "name": "Acceptance of Delivery",
-                "confidence": 0.8
-            },
-            {
-                "name": "Copyright License",
-                "confidence": 0.7
-            },
-            {
-                "name": "Demand Forecast",
-                "confidence": 0.6
-            },
-            {
-                "name": "Eat Apples",
-                "confidence": 0.5
-            },
-        ]
-    }
-
-
-
-class suggest_best_practice_body(BaseModel):
-    text: str = example_text
-
-
-
-@router.post("/model/suggestion", tags = ["Predict"], status_code=status.HTTP_200_OK)
-def suggest_best_practice(data: suggest_best_practice_body):
-    return {
-        "predict": "Not Finish Yet."
-    }
